+++ src/Smartstore.Core/Migrations/SmartDbContextDataSeeder.cs
@@ -46,6 +46,8 @@ namespace Smartstore.Core.Data.Migrations
 
             builder.AddOrUpdate("Admin.Configuration.Settings.Order.GiftCards_Activated.Hint")
                 .Value("de", "Legt den Auftragsstatus einer Bestellung fest, bei dem in der Bestellung enthaltene Geschenkgutscheine automatisch aktiviert werden.");
+
+            builder.Delete("Admin.Configuration.Currencies.GetLiveRates");
         }
 
         /// <summary>